requirements.txt
aniso8601==9.0.1
aspose-cells==24.9.0
blinker==1.7.0
boto3==1.35.32
botocore==1.35.32
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
copulas==0.11.1
ctgan==0.10.1
deepecho==0.6.0
dnspython==2.6.1
docutils==0.21.2
docx==0.2.4
email_validator==2.2.0
et-xmlfile==1.1.0
Faker==30.1.0
filelock==3.16.1
Flask==3.0.2
Flask-Cors==5.0.0
Flask-Login==0.6.3
Flask-Principal==0.4.0
Flask-RESTful==0.3.10
Flask-Security-Too==5.5.2
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
fsspec==2024.9.0
graphviz==0.20.3
greenlet==3.0.3
gunicorn==23.0.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
itsdangerous==2.1.2
Jinja2==3.1.3
jmespath==1.0.1
jni==1.0.0b22
joblib==1.4.2
JPype1==1.5.0
jvm==0.5.0b7
lxml==5.3.0
MarkupSafe==2.1.5
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
passlib==1.7.4
pillow==10.4.0
pkg-about==1.2.0
platformdirs==4.3.6
plotly==5.24.1
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-magic==0.4.27
pytz==2024.1
PyYAML==6.0.2
rdt==1.12.4
s3transfer==0.10.2
scikit-learn==1.5.2
scipy==1.14.1
sdmetrics==0.16.0
sdv==1.17.0
setuptools==75.1.0
six==1.16.0
SQLAlchemy==2.0.28
sympy==1.13.3
tenacity==9.0.0
threadpoolctl==3.5.0
torch==2.4.1
tqdm==4.66.5
typing_extensions==4.10.0
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.0.1
WTForms==3.1.2
zipp==3.20.2
